created new drag animation for carousel
