--- src/components/ui/FramerCarousel.tsx
@@ -1,173 +1,0 @@
-import { MouseEvent as ReactMouseEvent, useRef, useState } from "react";
-import Link from "next/link";
-import { motion, useMotionValue, useSpring, type PanInfo } from "framer-motion";
-import { MoveLeft, MoveRight } from "lucide-react";
-
-import { cn } from "@/lib/utils";
-
-const START_INDEX = 1;
-const DRAG_THRESHOLD = 150;
-const FALLBACK_WIDTH = 509;
-
-const CURSOR_SIZE = 80;
-
-interface FramerCarouselProps {
-  carousel: string[];
-}
-
-const FramerCarousel: React.FC<FramerCarouselProps> = ({ carousel = [] }) => {
-  const containerRef = useRef<HTMLUListElement>(null);
-  const itemsRef = useRef<(HTMLLIElement | null)[]>([]);
-  const [activeSlide, setActiveSlide] = useState(START_INDEX);
-  const canScrollPrev = activeSlide > 0;
-  const canScrollNext = activeSlide < carousel.length - 1;
-  const offsetX = useMotionValue(0);
-  const animatedX = useSpring(offsetX, {
-    damping: 20,
-    stiffness: 150,
-  });
-
-  const [isDragging, setIsDragging] = useState(false);
-
-  function handleDragSnap(
-    _: MouseEvent,
-    { offset: { x: dragOffset } }: PanInfo
-  ) {
-    //reset drag state
-    setIsDragging(false);
-    containerRef.current?.removeAttribute("data-dragging");
-
-    //stop drag animation (rest velocity)
-    animatedX.stop();
-
-    const currentOffset = offsetX.get();
-
-    //snap back if not dragged far enough or if at the start/end of the list
-    if (
-      Math.abs(dragOffset) < DRAG_THRESHOLD ||
-      (!canScrollPrev && dragOffset > 0) ||
-      (!canScrollNext && dragOffset < 0)
-    ) {
-      animatedX.set(currentOffset);
-      return;
-    }
-    let offsetWidth = 0;
-
-    for (
-      let i = activeSlide;
-      dragOffset > 0 ? i >= 0 : i < itemsRef.current.length;
-      dragOffset > 0 ? i-- : i++
-    ) {
-      const item = itemsRef.current[i];
-      if (item === null) continue;
-      const itemOffset = item.offsetWidth;
-
-      const prevItemWidth =
-        itemsRef.current[i - 1]?.offsetWidth ?? FALLBACK_WIDTH;
-      const nextItemWidth =
-        itemsRef.current[i + 1]?.offsetWidth ?? FALLBACK_WIDTH;
-
-      if (
-        (dragOffset > 0 && //dragging left
-          dragOffset > offsetWidth + itemOffset && //dragged past item
-          i > 1) || //not the first/second item
-        (dragOffset < 0 && //dragging right
-          dragOffset < offsetWidth + -itemOffset && //dragged past item
-          i < itemsRef.current.length - 2) //not the last/second to last item
-      ) {
-        dragOffset > 0
-          ? (offsetWidth += prevItemWidth)
-          : (offsetWidth -= nextItemWidth);
-        continue;
-      }
-      if (dragOffset > 0) {
-        //prev
-        offsetX.set(currentOffset + offsetWidth + prevItemWidth);
-        setActiveSlide(i - 1);
-      } else {
-        //next
-        offsetX.set(currentOffset + offsetWidth - nextItemWidth);
-        setActiveSlide(i + 1);
-      }
-      break;
-    }
-  }
-
-  function scrollPrev() {
-    //prevent scrolling past first item
-    if (!canScrollPrev) return;
-
-    const nextWidth = itemsRef.current
-      .at(activeSlide - 1)
-      ?.getBoundingClientRect().width;
-    if (nextWidth === undefined) return;
-    offsetX.set(offsetX.get() + nextWidth);
-
-    setActiveSlide((prev) => prev - 1);
-  }
-
-  function scrollNext() {
-    // prevent scrolling past last item
-    if (!canScrollNext) return;
-
-    const nextWidth = itemsRef.current
-      .at(activeSlide + 1)
-      ?.getBoundingClientRect().width;
-    if (nextWidth === undefined) return;
-    offsetX.set(offsetX.get() - nextWidth);
-
-    setActiveSlide((prev) => prev + 1);
-  }
-
-  const [hoverType, setHoverType] = useState<"prev" | "next" | "click" | null>(
-    null
-  );
-
-  const mouseX = useMotionValue(0);
-  const mouseY = useMotionValue(0);
-  const animatedHoverX = useSpring(mouseX, {
-    damping: 20,
-    stiffness: 400,
-    mass: 0.1,
-  });
-  const animatedHoverY = useSpring(mouseY, {
-    damping: 20,
-    stiffness: 400,
-    mass: 0.1,
-  });
-
-  function navButtonHover({
-    currentTarget,
-    clientX,
-    clientY,
-  }: ReactMouseEvent<HTMLButtonElement, MouseEvent>) {
-    const parent = currentTarget.offsetParent;
-    if (!parent) return;
-    const { left: parentLeft, top: parentTop } = parent.getBoundingClientRect();
-
-    const { left, top, width, height } = currentTarget.getBoundingClientRect();
-    const centerX = left + width / 2;
-    const centerY = top + height / 2;
-
-    const offsetFromCenterX = clientX - centerX;
-    const offsetFromCenterY = clientY - centerY;
-
-    mouseX.set(left - parentLeft + offsetFromCenterX / 4);
-    mouseY.set(top - parentTop + offsetFromCenterY / 4);
-  }
-
-  function disableDragClick(e: ReactMouseEvent<HTMLAnchorElement>) {
-    if (isDragging) {
-      e.preventDefault();
-      e.stopPropagation();
-    }
-  }
-
-  return (
-    <>
-      <div></div>
-    </>
-  );
-};
-
-export default FramerCarousel;
